PHP: Add unit testing workflow
This workflow runs PHP unit tests on the specified PHP version with optional
PHP extensions.  It sets up PHP, installs code dependencies, and runs PHPUnit
as a testing tool.  Additional dependencies variables can be cloned as
specified in the inputs.  The phpunit command uses a bootstrap file if the
PHPUNIT_BOOTSTRAP environment variable is set.  PHPUnit problem matchers are
added to improve error reporting in the workflow logs.